# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "bile-mo API",
        "description": "bile-mo is a web service offering a selection of mobile phones",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/bile-mo-api/v1",
            "description": "bile-mo API main server"
        }
    ],
    "paths": {
...
        "/customers": {
            "get": {
                "tags": [
                    "Customers management"
                ],
                "summary": "Returns list of customers",
                "operationId": "c992d6a1c8c2a6a9b7a4d958e62a79b0",
                "responses": {
                    "200": {
                        "description": "HTTP_OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Ok",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CustomerDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers management"
                ],
                "summary": "Create a new customer",
                "operationId": "4b9978cf561413798b854dcc3ca794b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "company": {},
                                    "enabled": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "HTTP_CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Created",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "HTTP_CONFLICT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Conflict",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Bad request",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/{customer_id}": {
            "get": {
                "tags": [
                    "Customers management"
                ],
                "summary": "Returns customer by id",
                "operationId": "3194ea66399f1cd3ad6548df7e1b3c51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTTP_OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Ok",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CustomerDTO"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Not found",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Customers management"
                ],
                "summary": "Updates a customer by id",
                "operationId": "64e4e4f64cb1a1a391352f8656525108",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "company": {},
                                    "enabled": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "HTTP_CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Created",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CustomerDTO"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "HTTP_CONFLICT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Conflict",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Bad request",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Not found",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customers management"
                ],
                "summary": "Deletes a customer by id",
                "operationId": "d9344752e61198b298352e36876b0034",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "customer ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "HTTP_NO_CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "No content",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Not found",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "code": {
                        "type": "string",
                        "nullable": false
                    },
                    "enabled": {
                        "type": "boolean",
                        "nullable": false
                    },
                    "company": {
                        "type": "string",
                        "nullable": false
                    },
                    "users": {
                        "type": "array"
                    }
                },
                "type": "object"
            },
...
        }
    }
}